@@ -4,17 +4,19 @@
   "description": "OpenTokRTC sample application to show off the OpenTok API and platform capabilities ",
   "main": "server.js",
   "engines": {
-    "node": "8.0.x",
-    "npm": "2.14.x"
+    "node": "12.19.x",
+    "npm": "6.14.x"
   },
   "scripts": {
     "start": "node server.js",
     "dev": "concurrently \"npm run start\" \"grunt clientDev\"",
     "postinstall": "bower install && grunt clientBuild",
     "test": "grunt test",
-    "regression-test": "wdio wdio.conf.js",
-    "lint": "./node_modules/.bin/eslint ./**/*.js",
-    "lint-auto-fix": "./node_modules/.bin/eslint ./**/*.js --fix"
+    "regression-test": "wdio run wdio.conf.js",
+    "karma-test": "karma start karma.conf.js",
+    "lint": "./node_modules/.bin/eslint .",
+    "lint-auto-fix": "./node_modules/.bin/eslint . --fix",
+    "clientBuild": "grunt clientBuild"
   },
   "repository": {
     "type": "git",
@@ -29,66 +31,79 @@
   "dependencies": {
     "accepts": "^1.3.7",
     "bower": "^1.6.5",
-    "cors": "^2.7.1",
+    "cors": "^2.8.5",
     "daemon": "^1.1.0",
-    "ejs": "^2.5.5",
-    "express": "^4.0.0",
-    "firebase": "^2.3.1",
-    "firebase-token-generator": "^2.0.0",
+    "ejs": "^3.1.5",
+    "emoji-picker-element": "^1.2.1",
+    "express": "^4.17.1",
     "fs": "0.0.1-security",
-    "geoip-lite": "^1.4.1",
+    "geoip-lite": "^1.4.2",
     "git-rev": "^0.2.1",
-    "google-auth-library": "^0.11.0",
-    "grunt": "^0.4.5",
-    "grunt-autoprefixer": "^2.2.0",
+    "google-auth-library": "^6.1.1",
+    "grunt": "^1.3.0",
+    "grunt-autoprefixer": "^3.0.4",
     "grunt-bower-task": "^0.5.0",
     "grunt-cli": "^1.3.2",
-    "grunt-contrib-clean": "^0.6.0",
-    "grunt-contrib-compress": "^1.4.3",
-    "grunt-contrib-connect": "^2.0.0",
-    "grunt-contrib-copy": "^0.8.1",
+    "grunt-contrib-clean": "^2.0.0",
+    "grunt-contrib-compress": "^1.6.0",
+    "grunt-contrib-concat": "^1.0.1",
+    "grunt-contrib-connect": "^3.0.0",
+    "grunt-contrib-copy": "^1.0.0",
     "grunt-contrib-less": "^2.0.0",
     "grunt-contrib-watch": "^1.1.0",
     "grunt-gitinfo": "^0.1.9",
     "grunt-html-build": "^0.7.1",
-    "grunt-karma": "^0.12.2",
-    "grunt-mocha-test": "^0.12.7",
+    "grunt-karma": "^4.0.0",
+    "grunt-mocha-test": "^0.13.3",
+    "grunt-terser": "^1.0.0",
     "haikunator": "^2.1.2",
-    "helmet": "^4.1.0",
-    "ioredis": "^4.6.3",
-    "lodash": "^4.17.19",
-    "opentok": "^2.3.0",
-    "qs": "^6.9.3",
-    "swagger-boilerplate": "^0.0.16",
-    "yamljs": "^0.2.10"
+    "helmet": "^4.1.1",
+    "http2": "^3.3.7",
+    "ioredis": "^4.17.3",
+    "lodash": "^4.17.21",
+    "opentok": "^2.10.0",
+    "qs": "^6.9.4",
+    "swagger-boilerplate": "^0.1.6",
+    "yamljs": "^0.3.0"
   },
   "devDependencies": {
-    "chai": "^3.3.0",
-    "chai-as-promised": "^5.2.0",
-    "concurrently": "^3.4.0",
-    "eslint": "^4.18.2",
-    "eslint-config-airbnb": "^15.0.1",
-    "eslint-config-airbnb-base": "^11.2.0",
-    "eslint-plugin-import": "^2.3.0",
-    "eslint-plugin-jsx-a11y": "^5.0.3",
-    "eslint-plugin-react": "^7.0.1",
-    "firefox-profile": "^1.0.3",
-    "geckodriver": "^1.19.1",
+    "@wdio/cli": "^6.6.7",
+    "@wdio/dot-reporter": "^6.6.6",
+    "@wdio/firefox-profile-service": "^6.6.0",
+    "@wdio/local-runner": "^6.6.7",
+    "@wdio/mocha-framework": "^6.6.7",
+    "@wdio/selenium-standalone-service": "^6.6.5",
+    "@wdio/spec-reporter": "^6.6.8",
+    "@wdio/sync": "^6.6.7",
+    "browserify": "^17.0.0",
+    "grunt-critical": "^2.0.0",
+    "chai": "^4.2.0",
+    "chai-as-promised": "^7.1.1",
+    "chromedriver": "^86.0.0",
+    "concurrently": "^5.3.0",
+    "eslint": "^7.12.1",
+    "eslint-config-airbnb": "^18.2.0",
+    "eslint-config-airbnb-base": "^14.2.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jsx-a11y": "^6.3.1",
+    "eslint-plugin-react": "^7.21.4",
+    "firefox-profile": "^4.0.0",
+    "geckodriver": "^1.20.0",
     "istanbul": "^0.4.5",
-    "karma": "^1.7.1",
-    "karma-chrome-launcher": "^0.2.3",
-    "karma-coverage": "^1.1.1",
-    "karma-firefox-launcher": "^1.0.1",
-    "karma-html2js-preprocessor": "^0.1.0",
-    "karma-mocha": "^1.3.0",
-    "mocha": "^6.2.2",
-    "mockfirebase": "^0.12.0",
-    "sinon": "^1.17.0",
-    "supertest": "^3.4.2",
-    "wdio": "^3.0.3",
-    "wdio-firefox-profile-service": "^0.1.0",
-    "wdio-mocha-framework": "^0.5.11",
-    "wdio-sauce-service": "^0.4.0",
-    "wdio-selenium-standalone-service": "0.0.9"
+    "karma": "^5.2.3",
+    "karma-browserify": "^7.0.0",
+    "karma-chai": "^0.1.0",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-coverage": "^2.0.3",
+    "karma-firefox-launcher": "^1.3.0",
+    "karma-html2js-preprocessor": "^1.1.0",
+    "karma-mocha": "^2.0.1",
+    "karma-sinon": "^1.0.5",
+    "mocha": "^8.1.3",
+    "object-assign": "^4.1.1",
+    "sinon": "^9.2.0",
+    "sinon-test": "^3.0.0",
+    "supertest": "^5.0.0",
+    "wdio-chromedriver-service": "^6.0.4"
   }
 }
