@@ -1,7 +1,4 @@
-'use strict';
-
-var GoogleAuth = require('google-auth-library');
-const Utils = require('swagger-boilerplate').Utils;
+const { OAuth2Client } = require('google-auth-library');
 
 /* eslint-disable class-methods-use-this */
 class DisabledGoogleAuthStategy {
@@ -19,14 +16,15 @@ class EnabledGoogleAuthStrategy {
   constructor(googleId, hostedDomain) {
     this.googleId = googleId;
     this.hostedDomain = hostedDomain;
-    this.auth = new GoogleAuth; // eslint-disable-line new-parens
-    this.client = new this.auth.OAuth2(googleId, '', '');
-    this.verifyIdTokenPromise = Utils.promisify(this.client.verifyIdToken, 1, this.client);
+    this.client = new OAuth2Client(googleId, '', '');
   }
 
   verifyIdToken(token) {
     return new Promise((resolve, reject) => {
-      this.verifyIdTokenPromise(token, this.googleId)
+      this.client.verifyIdToken({
+        idToken: token,
+        audience: this.googleId,
+      })
         .then((login) => {
           const payload = login.getPayload();
           if (this.hostedDomain && (this.hostedDomain !== payload.hd)) {
@@ -34,12 +32,12 @@ class EnabledGoogleAuthStrategy {
           }
           resolve();
         })
-        .catch(err => reject(err));
+        .catch((err) => reject(err));
     });
   }
 }
 
 module.exports = {
   DisabledGoogleAuthStategy,
-  EnabledGoogleAuthStrategy
+  EnabledGoogleAuthStrategy,
 };
