@@ -1,7 +1,7 @@
 var chai = require('chai');
 var request = require('supertest');
 
-var expect = chai.expect;
+var { expect } = chai;
 
 const TEST_LOG_LEVEL = 0;
 
@@ -18,25 +18,21 @@ describe('OpenTokRTC server', () => {
     MockOpentok = require('../mocks/mock_opentok.js');
     process.env.TEMPLATING_SECRET = '123456';
 
-    // Note that we're not actually testing anything that uses Firebase here. They'll
-    // have their own unit tests. We're only avoiding using it at all.
     var mocks = {
       Opentok: MockOpentok,
-      Firebase: require('../mocks/mock_firebase'),
     };
 
     // Note that this actually executes on the level where the Grunt file is
     // So that's what '.' is. OTOH, the requires are relative to *this* file.
     // Yep, I don't like that either. Nope, I can't do anything about that.
     var YAML = require('yamljs');
-    var loadYAML =
-         apiFile => new Promise((resolve, reject) => {
-           try {
-             YAML.load(apiFile, resolve);
-           } catch (e) {
-             reject(e);
-           }
-         });
+    var loadYAML = (apiFile) => new Promise((resolve, reject) => {
+      try {
+        YAML.load(apiFile, resolve);
+      } catch (e) {
+        reject(e);
+      }
+    });
     loadYAML('./api.yml').then((apiSpec) => {
       app = (require('swagger-boilerplate').App)({
         modulePath: __dirname + '/../../server/', // eslint-disable-line no-path-concat
@@ -59,8 +55,8 @@ describe('OpenTokRTC server', () => {
     var aObject = aRes.body;
     for (var i = 0, l = aAttributes.length; i < l; i++) {
       if (!aObject[aAttributes[i]]) {
-        throw new Error('Missing required attribute: ' + aAttributes[i] +
-                        ' in ' + JSON.stringify(aObject));
+        throw new Error('Missing required attribute: ' + aAttributes[i]
+                        + ' in ' + JSON.stringify(aObject));
       }
     }
     return undefined;
@@ -129,14 +125,6 @@ describe('OpenTokRTC server', () => {
       .expect(200, done);
   });
 
-  it('GET /room/:roomName?template=room', (done) => {
-    request(app)
-      .get('/room/roomName?template=room')
-      .set('Accept', 'text/html')
-      .expect('X-XSS-Protection', '1; mode=block')
-      .expect(200, done);
-  });
-
   it('GET /room/:roomName?template=unknownTemplate without auth, return default', (done) => {
     request(app)
       .get('/room/roomName?template=unknownTemplate')
@@ -148,20 +136,20 @@ describe('OpenTokRTC server', () => {
   it('GET /room/:roomName?template=unknownTemplate&template_auth=1234 invalid auth',
     (done) => {
       request(app)
-      .get('/room/roomName?template=unknownTemplate&template_auth=1234')
-      .set('Accept', 'text/html')
-      .expect('Content-Type', new RegExp('text/html'))
-      .expect(200, done);
+        .get('/room/roomName?template=unknownTemplate&template_auth=1234')
+        .set('Accept', 'text/html')
+        .expect('Content-Type', new RegExp('text/html'))
+        .expect(200, done);
     });
 
   it('GET /room/:roomName?template=unknownTemplate&template_auth=123456 valid auth',
     (done) => {
       request(app)
-      .get('/room/roomName?template=unknownTemplate&template_auth=123456')
-      .set('Accept', 'application/json')
-      .expect('Content-Type', new RegExp('application/json'))
-      .expect(checkForAttributes.bind(undefined, ReturnError))
-      .expect(400, done);
+        .get('/room/roomName?template=unknownTemplate&template_auth=123456')
+        .set('Accept', 'application/json')
+        .expect('Content-Type', new RegExp('application/json'))
+        .expect(checkForAttributes.bind(undefined, ReturnError))
+        .expect(400, done);
     });
 
   it('POST /room/:roomName/archive should allow composite archiving', (done) => {
@@ -198,6 +186,7 @@ describe('OpenTokRTC server', () => {
   it('DELETE /archive/:archiveId', (done) => {
     request(app)
       .delete('/archive/12345')
+      .set('referer', 'https://example.com/room/unitTestRoom')
       .expect(405, done);
   });
 
@@ -216,16 +205,15 @@ describe('OpenTokRTC server', () => {
 
   it('GET /server/health should return 400 and expected values', (done) => {
     request(app)
-    .get('/server/health')
-    .expect(400)
-    .then((response) => {
-      console.log(response.body);
-      expect(response.body.name).to.equal('opentok-rtc');
-      expect(response.body.version).to.be.a.string;
-      expect(response.body.gitHash).to.be.a.string;
-      expect(response.body.status).to.equal('fail');
-      done();
-    });
+      .get('/server/health')
+      .expect(400)
+      .then((response) => {
+        expect(response.body.name).to.equal('opentok-rtc');
+        expect(response.body.version).to.be.a.string;
+        expect(response.body.gitHash).to.be.a.string;
+        expect(response.body.status).to.equal('fail');
+        done();
+      });
   });
   it('GET /thanks should return post meeting screen', (done) => {
     request(app)
