@@ -1,5 +1,9 @@
-var assert = chai.assert;
-var expect = chai.expect;
+var sinonTest = require('sinon-test');
+
+var test = sinonTest(sinon);
+sinon.test = test;
+var { assert } = chai;
+var { expect } = chai;
 var should = chai.should();
 
 describe('Utils', () => {
@@ -116,10 +120,12 @@ describe('Utils', () => {
         output: '?key=value&key2&key3=value3',
       },
       {
-        input: { key: 'value',
+        input: {
+          key: 'value',
           key2: undefined,
           key3: ['value3', 'value4', 'value5'],
-          key4: undefined },
+          key4: undefined
+        },
         output: '?key=value&key2&key3=value3&key3=value4&key3=value5&key4',
       },
     ];
@@ -246,8 +252,8 @@ describe('Utils', () => {
     ];
 
     useCases.forEach((useCase) => {
-      it('should generate as params' + JSON.stringify(useCase.output) + ' for ' +
-         useCase.input, () => {
+      it('should generate as params' + JSON.stringify(useCase.output) + ' for '
+         + useCase.input, () => {
         var result = Utils.parseSearch(useCase.input);
         expect(result.params).to.be.deep.equal(useCase.output);
         results.push(result);
@@ -258,13 +264,13 @@ describe('Utils', () => {
       it('should always return a getFirstValue method on the object\'\'', sinon.test(() => {
         results.forEach((aResult) => {
           expect(aResult.getFirstValue).to.exist;
-          expect(aResult.getFirstValue).to.be.a.function;
+          expect(aResult.getFirstValue).to.be.a('function');
         });
       }));
 
       results.forEach((result, index) => {
-        it('should return ' + useCases[index].getFirst.output + ' when called with ' +
-           useCases[index].getFirst.input + 'on the use case #' + index, () => {
+        it('should return ' + useCases[index].getFirst.output + ' when called with '
+           + useCases[index].getFirst.input + 'on the use case #' + index, () => {
           var useCase = useCases[index].getFirst;
           expect(result.getFirstValue(useCase.input)).to.be.equal(useCase.output);
         });
