@@ -1,5 +1,10 @@
-var assert = chai.assert;
-var expect = chai.expect;
+var sinonTest = require('sinon-test');
+
+var test = sinonTest(sinon);
+sinon.test = test;
+
+var { assert } = chai;
+var { expect } = chai;
 var should = chai.should();
 
 describe('FeedbackView', () => {
