@@ -1,6 +1,6 @@
 /* global Chat, TextProcessor, otHelper */
 
-!(exports => {
+!((exports) => {
   let usrId;
 
   let closeChatBtn;
@@ -10,6 +10,8 @@
   let chatContainer;
   let chatContent;
   let chatForm;
+  let emojiPicker;
+  let toggleEmojiBtn;
   const chatParticipants = [];
 
   let _visibilityChanging = Promise.resolve();
@@ -17,9 +19,7 @@
   function isMobile() { return typeof window.orientation !== 'undefined'; }
 
   function isVisible() {
-    return _visibilityChanging.then(() => {
-      return Chat.visible;
-    });
+    return _visibilityChanging.then(() => Chat.visible);
   }
 
   function setVisibility(isVisible) {
@@ -44,36 +44,36 @@
       incomingMessage: {
         name: 'chatController:incomingMessage',
         handler(evt) {
-          const data = evt.detail.data;
+          const { data } = evt.detail;
           insertChatLine(data);
-          isVisible().then(visible => {
+          isVisible().then((visible) => {
             if (!visible) {
               Utils.sendEvent('chatView:unreadMessage', { data });
             }
           });
-        }
+        },
       },
       presenceEvent: {
         name: 'chatController:presenceEvent',
         handler(evt) {
           insertChatEvent(evt.detail);
-        }
+        },
       },
       messageDelivered: {
         name: 'chatController:messageDelivered',
         handler() {
           chatMsgInput.value = '';
-        }
+        },
       },
       chatVisibility: {
         name: 'roomView:chatVisibility',
         handler(evt) {
           _visibilityChanging = setVisibility(evt.detail);
         },
-        couldBeChanged: true
-      }
+        couldBeChanged: true,
+      },
     };
-    Array.isArray(configuredEvts) && configuredEvts.forEach(aEvt => {
+    Array.isArray(configuredEvts) && configuredEvts.forEach((aEvt) => {
       const event = eventHandlers[aEvt.type];
       event && event.couldBeChanged && (event.name = aEvt.name);
     });
@@ -89,9 +89,23 @@
     chatContainer = chatWndElem.querySelector('#chatMsgs');
     chatContent = chatContainer.querySelector('ul');
     chatForm = chatWndElem.querySelector('#chatForm');
+    emojiPicker = document.querySelector('emoji-picker');
+    toggleEmojiBtn = chatWndElem.querySelector('#addEmoji');
   }
 
-  const onSendClicked = evt => {
+  const onEmojiClicked = (event) => {
+    chatMsgInput.value += (` ${event.detail.unicode} `);
+  };
+
+  const toggleEmojiView = (evt) => {
+    evt.preventDefault();
+    if (emojiPicker.style.display === 'none') {
+      emojiPicker.style.display = 'block';
+    } else {
+      emojiPicker.style.display = 'none';
+    }
+  };
+  const onSendClicked = (evt) => {
     evt.preventDefault();
     if (!chatMsgInput.value.trim().length) {
       return;
@@ -104,8 +118,11 @@
     Utils.sendEvent('chatView:outgoingMessage', {
       sender: usrId,
       time: Utils.getCurrentTime(),
-      text: chatMsgInput.value.trim()
+      text: chatMsgInput.value.trim(),
     });
+    if (emojiPicker) {
+      emojiPicker.style.display = 'none';
+    }
   };
 
   const onKeyPress = ((myfield, evt) => {
@@ -127,12 +144,12 @@
     return true;
   }).bind(undefined, chatMsgInput);
 
-  const onSubmit = evt => {
+  const onSubmit = (evt) => {
     evt.preventDefault();
     return false;
   };
 
-  const onClose = evt => {
+  const onClose = (evt) => {
     evt.preventDefault();
     evt.stopImmediatePropagation();
     _visibilityChanging = setVisibility(false);
@@ -142,7 +159,7 @@
     Chat.isCollapsed() ? Chat.expand() : Chat.collapse();
   };
 
-  const onDrop = evt => {
+  const onDrop = (evt) => {
     evt.preventDefault();
     evt.stopPropagation();
     return false;
@@ -157,6 +174,10 @@
     closeChatBtn.addEventListener('click', onClose);
     headerChat.addEventListener('click', onToggle);
     sendMsgBtn.addEventListener('click', onSendClicked);
+    if (emojiPicker) {
+      emojiPicker.addEventListener('emoji-click', onEmojiClicked);
+      toggleEmojiBtn.addEventListener('click', toggleEmojiView);
+    }
   }
 
   function removeHandlers() {
@@ -165,6 +186,10 @@
     headerChat.removeEventListener('click', onToggle);
     sendMsgBtn.removeEventListener('click', onSendClicked);
     chatForm.removeEventListener('drop', onDrop);
+    if (emojiPicker) {
+      emojiPicker.removeEventListener('emoji-click', onEmojiClicked);
+      toggleEmojiBtn.removeEventListener('click', toggleEmojiView);
+    }
   }
 
   function insertChatEvent(data) {
@@ -180,7 +205,7 @@
   function insertText(elemRoot, text) {
     const txtElems = TextProcessor.parse(text);
     const targetElem = HTMLElems.createElementAt(elemRoot, 'p');
-    txtElems.forEach(node => {
+    txtElems.forEach((node) => {
       switch (node.type) {
         case TextProcessor.TYPE.URL:
           HTMLElems.createElementAt(targetElem, 'a',
@@ -224,7 +249,7 @@
   function init(aUsrId, configuredEvts) {
     return LazyLoader.dependencyLoad([
       '/js/helpers/textProcessor.js',
-      '/js/components/chat.js'
+      '/js/components/chat.js',
     ]).then(() => {
       initHTMLElements();
       usrId = aUsrId;
@@ -234,7 +259,7 @@
   }
 
   const ChatView = {
-    init
+    init,
   };
 
   exports.ChatView = ChatView;
