@@ -15,6 +15,8 @@
     let getStatsIntervalId;
     let testTimeoutId;
     let currentStats;
+    let testRunning;
+    let publishing;
 
     const testStreamingCapability = (subscriber, callback) => {
       performQualityTest({subscriber, timeout: TEST_TIMEOUT_MS}, (error, results) => {
@@ -95,6 +97,7 @@
     };
 
     this.startNetworkTest = callback => {
+      testRunning = true;
       publisher.publishVideo(true);
       const callbacks = {
         onInitPublisher: function onInitPublisher(error) {
@@ -111,8 +114,13 @@
             console.error('Could not publish video.', error);
             return;
           }
+          publishing = true;
+          if (!testRunning) {
+            session.unpublish(publisher);
+            return session.disconnect();
+          }
 
-          subscriber = session.subscribe(
+          return subscriber = session.subscribe(
             publisher.stream,
             subscriberEl,
             {
@@ -124,13 +132,18 @@
         },
 
         cleanup() {
+          subscriber && session.unsubscribe(subscriber);
+          publisher && session.unpublish(publisher);
           session.disconnect();
         },
 
         onSubscribe: function onSubscribe(error, subscriber) {
-          if (error) {
-            console.error('Could not subscribe to video.', error);
+          if (!testRunning) {
             return;
+          }
+
+          if (error) {
+            return console.error('Could not subscribe to video.', error);
           }
 
           testStreamingCapability(subscriber, (error, result) => {
@@ -140,10 +153,13 @@
         },
 
         onConnect: function onConnect(error) {
+          if (!testRunning) {
+            return;
+          }
           if (error) {
             console.error('Could not connect to OpenTok.', error);
           }
-          session.publish(publisher, callbacks.onPublish);
+          return session.publish(publisher, callbacks.onPublish);
         }
       };
 
@@ -159,6 +175,9 @@
 
       callbacks.onInitPublisher();
 
+      if (!testRunning) {
+        return;
+      }
       // This publisher uses the default resolution (640x480 pixels) and frame rate (30fps).
       // For other resoultions you may need to adjust the bandwidth conditions in
       // testStreamingCapability().
@@ -167,18 +186,24 @@
       });
 
       session = OT.initSession(options.apiKey, options.sessionId);
-      session.connect(options.token, callbacks.onConnect);
+      return session.connect(options.token, callbacks.onConnect);
     };
     
     this.stopTest = () => {
-      bandwidthCalculator && bandwidthCalculator.stop();
-      try {
-        session.unpublish(publisher);
-        session.disconnect();
-      } catch(error) {
-        // Probably not connected yet.
+      if (!testRunning) {
+        return;
       }
-      publisher.destroy();
+      testRunning = false;
+      bandwidthCalculator && bandwidthCalculator.stop();
+      if (publishing) {
+        try {
+          session.unpublish(publisher);
+          subscriber && session.unsubscribe(subscriber);
+          session.disconnect();
+        } catch(error) {
+          // Probably not connected yet.
+        }
+      }
     }
 
     // Helpers
