@@ -13,7 +13,6 @@
       if (aDataType) {
         // Note that this requires
         xhr.setRequestHeader('Content-Type', aDataType);
-        aData && xhr.setRequestHeader('Content-Length', aData.length);
       }
 
       xhr.onload = aEvt => {
@@ -103,20 +102,6 @@
     return sendXHR('DELETE', `${server}/archive/${id}`);
   }
 
-  function saveConnection(connection, sessionId) {
-    return sendXHR('POST', `${server}/saveConnection/`, JSON.stringify({
-      connection,
-      sessionId
-    }), 'application/json');
-  }
-
-  function deleteConnection(connection, sessionId) {
-    return sendXHR('POST', `${server}/deleteConnection/`, JSON.stringify({
-      connection,
-      sessionId
-    }), 'application/json');
-  }
-
   const Request = {
     getRoomInfo,
     getRoomRawInfo,
@@ -125,9 +110,7 @@
     dialOut,
     hangUp,
     deleteArchive,
-    sendXHR,
-    saveConnection,
-    deleteConnection
+    sendXHR
   };
 
   exports.Request = Request;
