@@ -1,22 +1,20 @@
-//  globals Firebase
-
-!(exports => {
-  const archives = null;
+!((exports) => {
+  let archives = null;
   const listeners = {};
 
   const archiveHandler = {
     archiveUpdates(evt) {
       const handlers = listeners.value;
       const archiveValues = Promise.resolve(evt.detail || {});
-      handlers && handlers.forEach(aHandler => {
+      handlers && handlers.forEach((aHandler) => {
         archiveValues.then(aHandler.method.bind(aHandler.context));
       });
-    }
+    },
   };
 
   function init() {
     const self = this;
-    return new Promise(resolve => {
+    return new Promise((resolve) => {
       Utils.addEventsHandlers('roomController:', archiveHandler, exports);
       resolve(self);
     });
@@ -37,7 +35,7 @@
     if (hd) {
       listeners[type].push({
         method: hd,
-        context
+        context,
       });
     }
   }
@@ -62,14 +60,17 @@
     return false;
   }
 
-  const FirebaseModel = {
+  const ArchivesEventsListener = {
     addEventListener,
     removeEventListener,
     init,
+    set archives(existingArchives) {
+      archives = existingArchives;
+    },
     get archives() {
       return archives;
-    }
+    },
   };
 
-  exports.FirebaseModel = FirebaseModel;
+  exports.ArchivesEventsListener = ArchivesEventsListener;
 })(this);
