@@ -1,14 +1,12 @@
 /* global PhoneNumberView */
-!(global => {
-  const init = () => {
-    return LazyLoader.load(
-      '/js/phoneNumberView.js'
-    ).then(() => {
-      PhoneNumberView.init();
-    });
-  };
+!((global) => {
+  const init = () => LazyLoader.load(
+    '/js/min/phoneNumberView.min.js',
+  ).then(() => {
+    PhoneNumberView.init();
+  });
 
   global.PhoneNumberController = {
-    init
+    init,
   };
 })(this);
